Se terminó de agregar validaciones para update

+++ src/main/java/com/sofkau/crudPerson/services/PersonServices.java
@@ -1,6 +1,7 @@
 package com.sofkau.crudPerson.services;
 
 import com.sofkau.crudPerson.config.exceptions.BadRequestException;
+import com.sofkau.crudPerson.config.exceptions.NotFoundException;
 import com.sofkau.crudPerson.entities.Person;
 import com.sofkau.crudPerson.repositories.PersonInterfaceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,6 +23,9 @@ public class PersonServices implements PersonServicesInterface{
 
     @Override
     public Optional<Person> listPersonById(int id) {
+        if (data.findById(id).isEmpty()){
+            throw new NotFoundException("the user that you want to delete was not found");
+        }
         return data.findById(id);
     }
 
@@ -38,11 +42,26 @@ public class PersonServices implements PersonServicesInterface{
 
     @Override
     public void deletePerson(int id) {
+        Optional res = data.findById(id);
+        if (res.isEmpty()){
+            throw new NotFoundException("the user that you want to delete was not found");
+        }
         data.deleteById(id);
     }
 
     @Override
     public Person updatePerson(Person person) {
+        Optional res = data.findById(person.getId());
+        if (res.isEmpty()){
+            throw new BadRequestException("the user that you want to Update was not found ");
+        }else {
+            if (person.getName() == null || person.getName().isEmpty()){
+                throw new BadRequestException("Invalid name");
+            }
+            if (person.getAge() == null || person.getAge().isEmpty()){
+                throw new BadRequestException("Invalid age");
+            }
+        }
         return data.save(person);
     }
 }
